+++ CHANGELOG.rst
@@ -19,11 +19,17 @@ Added
 * Input type for **start_tracker_thread** method for param **state_function** is changed from **Callable** to **str**
 * **timeout_decorator** and **error_propagation_decorator** added for implementing timeout and error propagation functionalities respectively
 
+[0.15.0]
+************
+* Updated Scan command interface to include scan_id as argument
+* EndScan command has been added in HelperDishLNDevice
+* **scanID** attribute has been introduced in HelperDishDevice
+
 
 Fixed
 -----
 
-[0.14.1]
+[0.15.1]
 ************
 
 * Updated **push_command_result** method from the HelperBaseDevice to take correct number of arguments